Remove durability class, integrated into tool
